+++ src/components/ui/logo.tsx
@@ -1,8 +1,9 @@
+import { cn } from "@/lib/utils";
 import Image from "@/components/ui/image";
 
-export function Logo() {
+export function Logo({ className }: { className?: string }) {
   return (
-    <div className="mx-auto mb-4 size-8 lg:mx-[inherit]">
+    <div className={cn("size-8 bg-white", className)}>
       <Image src={"/grbpwr-logo.webp"} alt="grpwr logo" aspectRatio="1/1" />
     </div>
   );
